fix IAM joining failing in some cases when a web identity was used as an AWS credential
When the Teleport service attempting to join a cluster using an IAM token
was running in FIPS mode in an AWS region without a FIPS STS endpoint,
and a web identity token was used as the AWS credential, the local AWS
region would be used to build the STS request. We had logic to detect this
and fallback to the `us-east-1` region instead, but that was done by
passing an additional config to `sts.New`. The config that the AWS
`session.Session` contains takes priority over all others. The `us-east-1`
region was initially used to build the STS request, but when calling
`Request.Sign` on the STS request the URL was changed to use the region
configured in the `session.Session`.

Instead, rebuild the `session.Session` and STS client `sts.STS` when the
region needs to be changed.