revert OIDC IdP CA to use RSA keys in new suites
I never actually updated any of our OIDC IdP code to support ECDSA, it
actually just doesn't support it at all.

It also goes against the OIDC spec to not support the RS256 algorithm for
id_token_signing_alg_values_supported, which we wouldn't be able to
support if we didn't have an RSA key.

I've already reverted the SAML IdP as well. We'll need a different
strategy to offer both RSA and ECDSA in the future.