[kube] Drop usage of `ProcessKubeCSR` (#46779)
* [kube] Drop usage of `ProcessKubeCSR`

This PR drops the usage of `ProcessKubeCSR` to ask Auth to sign a certificate on behalf of a user when running `tsh request search --kind=pod,...`.
Previously, we used `ProcessKubeCSR` to sign a new certificate but given that Kubernetes Access supports impersonation, this PR levarages it.

This also allows reusing the same kubernetes client for multiple calls.

* add deprecation notices to ProcessKubeCSR

* don't enforce kube api to be enabled