drop `types.PluginsV1` from audit log payload (#46830)
Okta plugin sets `plugin.spec.settings` which is a one of message and our audit log incorrectly uses `json.Marshal/Unmarshal` to marshal and Unmarshal objects from json.

When unmarshaling the oneof, json parser fails because it can't decode the correct message.

This PR changes the type to use a `google.protobuf.Struct` so it can hold any static data.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>