[kube] support tunneled SPDY protocol for outgoing connections
This PR extends Teleport Kuberetes Access to be able to use the new
tunneled SPDY protocol for portforward sessions.

Tunneled SPDY protocol consists in using SPDY pings and frames on top of
a websocket connections. This new protocol was introduced in Kubernetes
1.31.0 because most comonly used HTTP proxies will drop support for SPDY
protocol in the near future given that it's outdated for more than 6
years.

This PR checks if the target of the connections - teleport component or
Kubernetes API - support the new functionality. If the target Teleport
Kubernetes Service or Teleport Proxy don't support the funcionatilty,
the forwarder fallsback to use native SPDY protocol.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>