Remove oxy connections limiter
lib/limiter composes a maximum simultaneous connections limiter
as well as a rate limiter. This commit replaces the connections
limiter from oxy with built-in code.

Note: this also introduces a bug fix that may change behavior.
Prior to this change, the connection limiter kept a separate
connection account for HTTP connections than it did for connections
managed manually with acquire/release. We no longer maintain separate
counts - all connections (HTTP or not) contribute to the number
of allowed connections.