Add a firestore key migration to Teleport 17 (#46472)
* Add a firestore key migration to Teleport 17

Throughout the lifespan of the Firestore backend, the type of the key parameter has evolved from `string`, to `[]byte`, to `backend.Key`, and eventually back to `[]byte`. With each new type, documents were added without migrating the existing ones. As a result, every time a new key type was introduced, an additional step was required during key iteration because Firestore maps each of these key types to a distinct database type.

- `backend.Key` is mapped to an Array in Firestore.
- `[]byte` is mapped to Bytes in Firestore.
- `string` is mapped to a String in Firestore.

When searching for a specific key, the Go type is converted into a proto type and encoded in a particular way. During the database's range query, all keys with a different type than the requested one are ignored. This allows us to retrieve only the relevant keys and convert them to `[]byte` by updating the database.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* force key to be array

* add rate limited bucket

* reduce limit

* add migration in its own goroutine and use optimistic locking

* add deprecation release

* retry migration every 12h (max)

* add migration tests

* fix typo

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>