Document usage event validation
Closes #46392

Clarify whether it is possible to independently validate Teleport's
usage event data by describing the usage event aggregation service. Note
that validation is not possible. To add this information, split up the
"Anonymized usage data" section of the Usage and Billing page, then add
a subsection related to validating reports.