Remove v15 redirects
We only publish a sitemap for the default version of the docs site, and
with a couple of aberrant exceptions, do not link to past versions of
the docs on the corporate site or in the product. As a result, it is
safe to remove redirects from non-default versions to get us under the
Vercel route limit.