Increase feature slider height for smallest screens (#47058)
In our empty state pages, we have this neat purple frame that slides
over each feature. This gets very slightly wonky if the feature text is
a bit too large. This just bumps up the height for the smallest screens
to give a bit more breathing room to the text.