[entraid] replace external identifier from UserPrincipalName (#5084)
This PR removes the identifier that indicates a user is a B2B guest in EntraID. This identifier is available in entraID but when SAML assertion happens it's not present.

In Teleport cause, the presence of `#EXT#` causes the user import to fail due to invalid characters in backend key.

Closes https://github.com/gravitational/access-graph/issues/1190

Source-commit: gravitational/teleport.e@ad71b1a290c2a60cad2d77af65b057d2b7237376
Source-pull-requests: gravitational/teleport.e#5084