tsh: Deduplicate the list of request IDs (#46780)
* tsh: Deduplicate the list of request IDs

It's possible to specify the same request multiple times with
tsh request create. The duplicates eventually get resolved before
we generate a certificate, but they do exist in the access request
resource. This can cause the size of the resource to exceed the
limits of a gRPC message and break listing.

* Limit access requests to a maximum of 300 resources

* Only run deduplication when the request is being created