[v16] Add new empty state for Bots page in the web UI (#47118)
* Add new empty state for Bots page in the web UI (#46773)

* Add new images

* Shared EmptyState components

* Add DisplayTile

* Add Bots EmptyState component

* Use EmptyState and update story/test

* Update images with manual shadow

design team removed shadows from images to make them consistant and
requested manual shadows added for images (and not tile scene)

* Fix fontsize

* Update word

* Add optional pagerPosition prop to device table (#47046)

This adds an optional prop that is required for the new empty state
tables (coming in an `e` PR)

* Increase feature slider height for smallest screens (#47058)

In our empty state pages, we have this neat purple frame that slides
over each feature. This gets very slightly wonky if the feature text is
a bit too large. This just bumps up the height for the smallest screens
to give a bit more breathing room to the text.