Edit the "Connecting Apps" guide (#47190)
Closes #12168

Edit the guide to enrolling web applications with a Teleport cluster to:
- Abide by our conventions for the structure of a how-to guide, e.g., to
  include H2s that indicate step numbers.
- Clarify some missing information from the original guide, such as
  where to start the Application Service as a service, how the
  Application Service and Proxy Service interact, and how labels work.

This change does not attempt to modify the conceptual information at the
end of the guide aside from placing more of it under the "Advanced
options" heading. This is because the product links to these sections in
messages.