Clarify the wording on the device trust authorization dialog (#47219)
"Cancel" was a bad name for the secondary button, because most
users expect cancel will close the modal and do nothing else.
The actual behavior is that the web session is launched either
way, and the user is only choosing whether or not to allow it
to be authorized for device trust.

Closes #45478