Temporarily remove default rate limiter (#5122)
We're updating Teleport's lib/limiter to remove the dependency
on gravitational/oxy (it transitively pulls in a bunch of mailgun
dependencies that we're trying to get rid of).

This commit is a temporary part of the refactor - by disabling
the construction of the default limiter we break teleport.e's
dependency on Teleport's lib/limiter, which will allow the
build to succeed while we merge the Teleport updates.

Source-commit: gravitational/teleport.e@e866be3f49099da9c79c7cbd876bbcd19b15a022
Source-pull-requests: gravitational/teleport.e#5122