Reconciler extensions for Identity Center integration
This patch extends the existing Reconciler in 2 ways:
- Makes the resource key type generic, and
- Adds an optional custom comparison function

The Identity Center integration interacts with (and reconciles) several different
resource types, and using the builtin `string` for all of the ID values is just
asking for the wrong ID to get put in the wrong service. To mitigate this, both
the Identity Center integration and SCIM use strongly-types IDs in all of their
services, making the use of an incorrect ID a compile-time error.

Adding the generic key type allows the Identity Center integration and SCIM
clients use the Reconciler without a heap of casting or map rebuilding.

Prior to this patch the only way to specify a custom comparision routine is to
implement `compare.IsEqual[T]` on the target resource types. The IC integration
uses resources are defined outside the IC integration package, and the new RFD153-
style resources forbid adding custom methods to protobuf-generated types in their
packages.

The addition of a custom comparison function allows a Reconciler to specify a
custom comparison routine that will work with RDF153-style resources without
having to deal with wrapper types and avoiding the same sort of casting/map-rebuilding
issues as the genric resource keys.

This patch also supplies backwards-compatibilty wrappers, in order to avoid having
to update the many, many Reconciler useage sites.