[v16] UserTasks: List by Integration service and WebAPIs (#47379)
* UserTasks: List by Integration service method (#47327)

This PR adds a new method that allows users to list UserTasks by
Integration.
Given this is our main use case when showing the UserTasks in the UI
(we'll be always showing UserTasks for a given integration), we created
a dedicated method in the gRPC.
We'll probably add more filters but they will always include the
integration (at least for the uses cases we have currently).

* WebAPI: list User Tasks and mark as resolved (#47329)

* WebAPI: list User Tasks and mark as resolved

This PR adds new endpoints for the WebAPI related to UserTasks:
- list by integration
- get user task detail
- mark user task as resolved

* remove leftovers

* remove eice issue types