Clean up user and login enrichment code for v17 (#47459)
The users service was originally slated to be removed in v16, but
out of an abundance of caution it was left around for another
release. Now that unsupported connections are rejected there should
be no way to attempt to modify a user with a client still attempting
to use th REST API or the legacy gPRC API. Additionally, the
calculation of logins in the Web UI should no longer need to be
derived locally. All leaf clusters connected to a v17 root, must
be running at least v16, which introduced the new enriched API.