Clarify State vs Status in SCIM provisioning records (#47424)
* Clarify State vs Status in SCIM provisioning records

The word `Status` was overloaded in the provisionin state records, conflating the
ideas of `Status` (the current collection of runtime data about the resource) and
`State` (which particular part of the resource lifecycle state machine the
resource is in).

This change updates the record definition to disambiguate State and Status. Also
adds documentation to the State values.

* buf linter appeasement