Propagate redirect_uri during device trust web auth (#47520)
In the web UI, if a user tries to access a page but they are
unauthenticated, after logging in, they are automatically redirected
back to the originally requested page. If the user has to authenticate
their session with device trust, this URI got swallowed and the user
would always end up at "/web". This PR passes the redirect through to
the authentication process.