UserNotifications: add UpsertGlobalNotification
This PR adds the UpsertGlobalNotification method.
It allows clients to create and update GlobalNotifications.

This is required for the UserTasks.
The goal is to let the user know that they have a pending user tasks for
a given integration.
Instead of creating multiple notifications, we create a single one with
a specific user name.
Everytime we try to create another notification for that integration, it
will just issue an Upsert and only update the expiration.