[aws] exclude session recordings from S3 sync in `teleport-renew-cert` (#47610)
This PR excludes `records` directory from the sync when renewing the letsencrypt certificate.
When doing the wildcard sync, `aws sync` downloads the `records` folder which contains audit logs for the cluster which causes failures because of no space left on the disk.
The certbot hook `teleport-upload-cert` doesn't use the `--delete` flag so the records are never purged from S3 on upload

Fixes #27884

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>