Simplify how-to guides (#47586)
Closes #38931

Remove "Static" and "Dynamic" configuration branching. Use the dynamic
approach unless the guide already shows readers how to use the static
approach elsewhere, e.g., using `teleport db configure create`.

For example, in the AWS OpenSearch guide, creating a Teleport
configuration is required for the guide anyway, so this change uses only
the static approach.

Ignore edge cases where the branching has some qualities that are unique
to a specific guide:

- `application-access/cloud-apis/google-cloud.mdx`: includes branching
  on authentication preference type.
- `application-access/guides/connecting-apps.mdx` because it explains
  syntax differences between the static and dynamic approaches.
- `enroll-aws-databases/aws-cross-account.mdx`: branches using three
  tabs, so editing requires another approach than the one this change
  applies.

This change edits only how-to guides, not conceptual guides, since the
goal of a how-to guide is to get the user to an end state as quickly as
possible.