build: Add release-ng-arm64-fips release target (#46930)
Add a release target to `build.assets/Makefile` to build the arm64-fips
release. This was missed before because I erroneously assumed we stopped
doing that due to rust build issues, but it was just that rdpclient was
no longer part of the arm64-fips build, not that we abandoned that build
altogether.