Make the Vale prose linter required
As of Vale 3.8.0, it is possible to set custom comment delimiters so an
author can temporarily disable a style rule within an MDX docs page. For
example, an author can ignore the `protocol-products` rule by including
this:

```text
{/* vale messaging.protocol-products = NO */}

This paragraph is about server access.

{/* vale messaging.protocol-products = YES */}
```

As a result, we can require the Vale prose linter without worrying about
false positives.

This change updates the version of Vale used by the linting job to
3.8.0.

It also makes the prose linting step part of the required Lint (Docs)
job. To do so, it overrides `yarn git-update` to ignore submodules,
checks out `gravitational/teleport` at the current branch, and copies
the `teleport` clone into the directory where the docs engine expects
submodules. It then edits the working directory of the Vale job.

Since the Vale job only ran on `pull_request` events, and not
`merge_group`, this change also adjusts all of `Lint (Docs)` to run on
`pull_request`.

Finally, this change edits the level of some Vale styles:
- `error` (which causes Vale to exit with a nonzero status code): styles
  where we want to ensure consistency/correctness, and where it is
  possible to disable the style using comments.
- `warning` (which causes Vale to print a message but exit with a zero
  status code): style rules that an author might need to ignore. This
  includes style rules with the `raw` scope, which cannot be disabled
  using comments.