use AWS ECS service per VPC for DB service (#47842)
Originally, single db enrollment deployed a DB service for an entire AWS region.
This behavior was recently changed to deploy the DB service for a
specific VPC in a region.
So now the AWS ECS service must be named after the VPC it is deployed to for
single db enrollment.