Prevent go-cmp from being imported in production code
Adds a new depguard rule to limit the use of go-cmp to test code.
There are a few notable exceptions for legacy code that relies on
go-cmp which should find other suitable means for comparisions.
However, in the interest in limiting new code from introducing more
production depenency on go-cmp, the existing violations are allowed
for the time being and should be refactored and removed from the list
in the future.