Enable `strict` TypeScript mode in `web/design` (#47850)
* Fix simple issues: missing types and null checks

* Unify `size` values between `Icon` and `Button`

`IconSize` allowed `extraLarge` and `ButtonSize` allowed `extra-large`. Now both use `extra-large`.

* Improve `DataTable` types

The most important change is that `onSort` callback now provides the entire `T` objects, instead of just the values for a given column. This has two advantages:
* it's easier to add proper types for it,
* allows the callsites to sort by a different field, so we will be able to remove `altSortKey` that was introduced just for this purpose.

* Rewrite `Indicator` to a functional component

* Make `Modal` and `Popover` strict

* Remove unneeded `PropTypes` from `LabelInput` and convert it to TS

* Make `StepSlider` strict

* Fix non-existing color

* Map `Event.time` to `Date` object to match its type

After adding `Cell` types, we are no longer able to render `time` directly since it has `Date` type. But how did it even work? It turned out that it was just a string, not a Date.
`toISOString` renders the cell in the same format as before.

* Enable `strict` for `web/design` 🎉

* Enable `noImplicitReturns`

* Make `dialogCss` non-optional

* Add `LabelStatePropsWithoutKind`

* Improve displaying `copyToClipboard` error

* Simplify `dateMatcher` and `dateTimeMatcher`

* `DataTable` related fixes

* Add a comment about non-null assertions

* Use `Theme` instead of `DefaultTheme`

* Return `null` instead of `undefined` from `dialogEl`

* Update `e`