Add SSO as MFA confirmation page (#5229)
This PR adds a route/page that can be redirected to after successfully
authenticated with SSO (for an MFA action). The purpose for the new page
is that we need a way for this page to communicate to whatever page
orginally requested the auth async so we can wait for the confirmation
from sso to perform the requested action.

Source-commit: gravitational/teleport.e@9af43d5756f1cccebac19126077585728dd1e751
Source-pull-requests: gravitational/teleport.e#5229