[v17] sync changelog to v16
The master CHANGELOG.md which has become the changelog for v17 has no
entries for the releases since 14.0.0. However, some links in the master
changelog have been updated to point to new docs locations on master/v17
so we cannot just wholesale replace the changelog.

Add the entries from 14.0.0 onwards from the v16 changelog and update
docs/cspell.json to add entries for new words from that part of the
changelog.