[v16] Create DiscoverEC2 User Tasks when Auto Discover fails on EC2 instances (#47618)
* Create DiscoverEC2 User Tasks for failed EC2 enrollments

This PR changes the DiscoveryService to start creating and updating
Discover EC2 User Tasks.

So, what are Discover EC2 User Tasks?
When users set up Auto Discover for EC2 Instances, they don't have a
good way of checking for issues on their configured matchers.

We created User Tasks as a way to warn Users that something's wrong.
Each User Task should describe an issue that happened and a way to fix
it.
This has potential to be used to report unexpected events trough the
whole system, which are not errors per se, but something the user should
take action in order to improve the situation.
In this case, we are creating a sub type of those tasks: DiscoverEC2.

From now on, when the DiscoveryService fails to auto-enroll an instance,
it will create a DiscoverEC2 User Task grouping all the failed instances
by the following props:
- integration
- issue type
- account id
- region

A follow up PR will also create notifications so that the user can
actually be notified on those User Tasks and take action.

* improve naming and split function into multiple methods

* rename task issues variable name

* fix context early cancelation

* UserTasks: Add SSM Document and Installer for DiscoverEC2 Issues (#47626)

This PR adds two new fields to the DiscoverEC2 User Task.
SSM Document used to install teleport
Teleport Installer script name used to install teleport

This can be used to let the user know which scripts were used.
For the SSM Document, users can then open it in webui.
For the Installer Script, users can manage it using `tctl` client.

* Add instance name to DiscoverEC2 User Task failed instances list (#47712)

* Add instance name to DiscoverEC2 User Task failed instances list

* use stdlib slices and maps

* remove 1.23+ methods from slices and maps packages