Use static licenses in tests (#5257)
Generating tests dynamically in tests brings in several additional
dependencies from the gravitational/license repo, which in turn
causes cfssl to show up as a direct dependency. While the long
term goal is to remove cfssl from gravitational/license, in the
short term the tests have been updated to read a set of pregenerated
licenses to allow the dependencies to be reduced.

Source-commit: gravitational/teleport.e@84f36a00ae432a6d8b08d6b84dd56a4970402bcf
Source-pull-requests: gravitational/teleport.e#5257