Simplify the docs CI pipeline
Closes #35796

**Only run the workflow on `pull_request`**. It is not strictly
necessary to run the workflow on `merge_group`, and doing so on
`pull_request` simplifies the branching logic we use in the workflow.

**Test the docs build using the `checkout` GitHub Actions workflow**,
rather than using the name of the branch. This way, we don't need custom
logic to identify the branch name. To get this to work, turn the
`git-update` script in `package.json` into a no-op script. This will
also make it easier to add the prose linting step to `Lint (Docs)` in
the future, since we pull the `gravitational/teleport` source outside of
the `yarn update-and-build` command.

**Combine the Vercel preview workflow and "test the docs build" step:**
Both of these run the docs build, so there is no need to separate the
build step from the Vercel preview. This saves around 8-10 minutes of
runner time per workflow run.