Add a timeout to Lint (Docs) dep installation (#47994)
For reasons that are unclear, the dependency installation step in the
Lint (Docs) GitHub Actions job occasionally takes an indefinite time to
complete, and must be canceled manually or after reaching the maximum
GitHub Actions runtime.

To prevent this situation, this change adds a timeout of 10 minutes to
the step that installs NodeJS dependencies for the Lint (Docs) job. This
step tends to take around 7 minutes to complete, so a 10-minute timeout
should be reasonable.