tooling: Fix build of plugins tools (#5271)
Fix the build of the plugins tooling to use Go 1.23 and to use
`slices.Collect()` around `maps.Keys()`. A recent change (#4792)
deprecated `golang.org/x/exp/maps` in favour of the built-in `maps`
package, but the plugins tooling has its own `go.mod` file and it was
not up to Go 1.23 as required for these changes.

Fix the module name while we're here - these tools were migrated from
the gravitational/teleport-plugins repository and had a "correct" module
line for that, but now it's in teleport.e, update the module line to
match. This didn't actually matter all that much as the only imports
were from within the module, but it does end up giving some misleading
output on errors, so fix it.

e.g.

    # github.com/gravitational/teleport-plugins/tooling/cmd/promote-terraform
    Error: cmd/promote-terraform/main.go:130:18: undefined: maps.Keys

This leads you to believe to is running code from the teleport-plugins
repo, not teleport.e.

Source-commit: gravitational/teleport.e@b23222d7c1c5a747f41a95fb98d15e0073f7cd99
Source-pull-requests: gravitational/teleport.e#5271